requirements.txt
certifi==2018.11.29
chardet==3.0.4
ddt==1.2.1
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4
mysql-connector-python==8.0.15
openpyxl==2.6.1
protobuf==3.7.0
PyMySQL==0.9.3
requests==2.21.0
six==1.12.0
urllib3==1.24.1
xlrd==1.2.0
